Get backups working
Still some UI stuff to sort out.

+++ functions/backup.functions.php
@@ -1,38 +1,37 @@
 <?php
 
-/**
- * Backup database and files
+/* Backup database and files
+ *
+ * Creates a temporary directory containing a copy of all files
+ * and a dump of the database. Then zip that up and delete the temporary files
+ *
+ * @uses hmbkp_backup_mysql
+ * @uses hmbkp_backup_files
+ * @uses hmbkp_delete_old_backups
  */
 function hmbkp_do_backup() {
 
+	global $hm_backup;
+ 
+	// Make sure it's possible to do a backup
+	if ( !hmbkp_possible() )
+		return false;
+ 
 	// Clean up any mess left by the last backup
 	hmbkp_cleanup();
-
-	$backup = new HMBackup();
 	
-	$backup->path = hmbkp_path();
-	$backup->files_only = hmbkp_get_files_only();
-	$backup->database_only = hmbkp_get_database_only();
-	
-	if ( defined( 'MYSQLDUMP_PATH' ) && MYSQLDUMP_PATH )
-		$backup->mysqldump_path = MYSQLDUMP_PATH;
-
-	if ( defined( 'ZIP_PATH' ) && ZIP_PATH )
-		$backup->mysqldump_path = ZIP_PATH;
-		
-	$backup->excludes = hmbkp_get_excludes();
-	
-	$backup->backup();
-
+	$hm_backup->backup();
+ 
 	// Email Backup
-	hmbkp_email_backup( $filepath );
-
+	hmbkp_email_backup( $hm_backup->archive_filepath() );
+ 
     hmbkp_set_status( __( 'Removing old backups', 'hmbkp' ) );
-
+ 
 	// Delete any old backup files
     hmbkp_delete_old_backups();
     
-    unlink( hmbkp_path() . '/.backup_running' );
+    if ( file_exists( hmbkp_path() . '/.backup_running' ) )
+	    unlink( hmbkp_path() . '/.backup_running' );
     
 	$file = hmbkp_path() . '/.backup_complete';
 	
@@ -42,8 +41,7 @@ function hmbkp_do_backup() {
 	fwrite( $handle, '' );
 	
 	fclose( $handle );
-
-
+ 
 }
 
 /**
@@ -119,15 +117,6 @@ function hmbkp_delete_backup( $file ) {
 }
 
 /**
- * Check if a backup is running
- *
- * @return bool
- */
-function hmbkp_is_in_progress() {
-	return file_exists( hmbkp_path() . '/.backup_running' );
-}
-
-/**
   * Email backup.
   *
   *	@param $file
@@ -173,43 +162,79 @@ function hmbkp_email_backup( $file ) {
 }
 
 /**
+ * Set the status of the running backup
+ * 
+ * @param string $message. (default: '')
+ * @return void
+ */
+function hmbkp_set_status( $message = '' ) {
+	
+	$file = hmbkp_path() . '/.backup_running';
+	
+	if ( !$handle = @fopen( $file, 'w' ) )
+		return false;
+		
+	if ( is_object( $message ) ) {
+		hm( debug_backtrace() );
+		exit;
+	}
+	
+	fwrite( $handle, $message );
+	
+	fclose( $handle );
+	
+}
+
+/**
+ * Get the status of the running backup
+ * 
+ * @return string
+ */
+function hmbkp_get_status() {
+	
+	if ( !file_exists( hmbkp_path() . '/.backup_running' ) )
+		return false;
+		
+	return file_get_contents( hmbkp_path() .'/.backup_running' );
+	
+}
+
+/**
  * Return an array of invalid custom exclude rules
  *
  * @return array
  */
 function hmbkp_invalid_custom_excludes() {
- 
-    $invalid_rules = array();
- 
-    // Check if any absolute path excludes actually exist
-    if ( $excludes = hmbkp_get_excludes() )
-     
-        foreach ( explode( ',', $excludes ) as $rule )
-            if ( ( $rule = trim( $rule ) ) && in_array( substr( $rule, 0, 1 ), array( '/', '\\' ) ) && !file_exists( $rule ) && ! file_exists( ABSPATH . $rule ) && ! file_exists( trailingslashit( ABSPATH ) . $rule ) )
-                $invalid_rules[] = $rule;
- 
-    return $invalid_rules;
- 
+
+	$invalid_rules = array();
+
+	// Check if any absolute path excludes actually exist
+	if ( $excludes = hmbkp_get_excludes() )
+	
+		foreach ( explode( ',', $excludes ) as $rule )
+			if ( ( $rule = trim( $rule ) ) && in_array( substr( $rule, 0, 1 ), array( '/', '\\' ) ) && !file_exists( $rule ) && ! file_exists( ABSPATH . $rule ) && ! file_exists( trailingslashit( ABSPATH ) . $rule ) )
+				$invalid_rules[] = $rule;
+
+	return array_filter( $invalid_rules );
+
 }
- 
+
 /**
  * Return an array of valid custom exclude rules
  *
  * @return array
  */
 function hmbkp_valid_custom_excludes() {
- 
-    $valid_rules = array();
- 
-    $excludes = hmbkp_get_excludes();
-     
-    if( ! $excludes )
-        return;
- 
-    $valid_rules = array_diff( explode( ',', $excludes ), hmbkp_invalid_custom_excludes() );
-     
- 
- 
-    return array_map( 'trim', $valid_rules );
- 
+
+	$valid_rules = array();
+
+	$excludes = hmbkp_get_excludes();
+	
+	if ( ! $excludes )
+		array();
+
+	$valid_rules = array_diff( explode( ',', $excludes ), hmbkp_invalid_custom_excludes() );
+
+	return array_filter( array_map( 'trim', $valid_rules ) );
+
 }
